Fixed #79. Ignore Email global scope if app is running in console.

+++ src/Models/Email.php
@@ -53,10 +53,12 @@ class Email extends Model
     protected static function booted(): void
     {
         static::addGlobalScope('teams', function (Builder $query) {
-            if (auth()->check() && Filament::getTenant()) {
-                $query->whereBelongsTo(auth()->user()?->teams);
-            } else {
-                $query->whereTeamId(null);
+            if (!app()->runningInConsole()) {
+                if (auth()->check() && Filament::getTenant()) {
+                    $query->whereBelongsTo(auth()->user()?->teams);
+                } else {
+                    $query->whereTeamId(null);
+                }
             }
         });
     }
@@ -66,9 +68,9 @@ class Email extends Model
         parent::boot();
 
         self::deleting(function ($record) {
+            $folderPath = "";
             $storageDisk = config('filament-email.attachments_disk', 'local');
-            $folderPath = null;
-            if (! empty($record->attachments)) {
+            if (!empty($record->attachments)) {
                 foreach ($record->attachments as $attachment) {
                     $filePath = Storage::disk($storageDisk)->path($attachment['path']);
                     if (empty($folderPath)) {
@@ -76,14 +78,13 @@ class Email extends Model
                         array_pop($parts);
                         $folderPath = implode(DIRECTORY_SEPARATOR, $parts);
                     }
-                    if (! Storage::directoryExists($folderPath) && Storage::disk($storageDisk)->exists($attachment['path'])) {
+                    if (Storage::disk($storageDisk)->exists($attachment['path'])) {
                         Storage::disk($storageDisk)->delete($attachment['path']);
                     }
                 }
             }
-
-            if (! empty($record->raw_body) && count(explode(DIRECTORY_SEPARATOR, $record->raw_body)) === 3) {
-                if (! Storage::disk($storageDisk)->directoryExists($record->raw_body) && Storage::disk($storageDisk)->exists($record->raw_body)) {
+            if (!empty($record->raw_body) && count(explode(DIRECTORY_SEPARATOR, $record->raw_body)) === 3) {
+                if (Storage::disk($storageDisk)->exists($record->raw_body)) {
                     if (empty($folderPath)) {
                         $parts = explode(DIRECTORY_SEPARATOR, $record->raw_body);
                         array_pop($parts);
@@ -91,9 +92,9 @@ class Email extends Model
                     }
                     Storage::disk($storageDisk)->delete($record->raw_body);
                 }
-                if (Storage::disk($storageDisk)->directoryExists($folderPath)) {
-                    Storage::disk($storageDisk)->deleteDirectory($folderPath);
-                }
+            }
+            if (!empty($folderPath) && Storage::disk($storageDisk)->directoryExists($folderPath)) {
+                Storage::disk($storageDisk)->deleteDirectory($folderPath);
             }
         });
     }
